pyproject.toml
[project]
name = "sih2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "faker>=37.6.0",
    "gunicorn>=23.0.0",
    "markdown>=3.9",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
]
